Formalize "full week" and "overlapping week" nomenclature
"Full week" - all the days of this week belong to the same parent value. For example 7 Jan 2024 to 13 Jan 2024 is a "full week" of January.

"Overlapping week" - a week in which one or more days belong to the parent value. Multiple parent values can produce return the same overlapping week. For example, December 2023's last overlapping week is the *same week* as January 2024's first overlapping week.